butteraugli: now compiles.

+++ output_image.go
@@ -6,8 +6,8 @@ package guetzli_patapon
 
 type OutputImage struct{}
 
-func (img *OutputImage) ToLinearRGB(rgb [][]float64) {
+func (img *OutputImage) ToLinearRGB(rgb [][]float32) {
 }
 
-func (img *OutputImage) ToLinearRGB_(xmin, ymin, xsize, ysize int, rgb [][]float64) {
+func (img *OutputImage) ToLinearRGB_(xmin, ymin, xsize, ysize int, rgb [][]float32) {
 }
